+++ storage_drivers/astrads/api/v1beta1/doc.go
@@ -1,5 +1,0 @@
-package v1beta1
-
-// Package v1beta1 is the v1beta1 version of the API.
-// +groupName=astrads.netapp.io
-// +kubebuilder:object:generate=true